Move abstract receiver to `receivers` package (for now)
